Fixed bugs - MassEditOperation was barfing when engineConfig was missing - When parsing JSON in streaming mode, get long instead of int and double instead of float so that we won't get overflow exception.
git-svn-id: http://google-refine.googlecode.com/svn/trunk@476 7d457c2a-affb-35e4-300a-418c747d4874

+++ src/main/java/com/metaweb/gridworks/model/Cell.java
@@ -130,9 +130,9 @@ public class Cell implements HasFields, Jsonizable {
                 if (token == JsonToken.VALUE_STRING) {
                     value = jp.getText();
                 } else if (token == JsonToken.VALUE_NUMBER_INT) {
-                    value = jp.getIntValue();
+                    value = jp.getLongValue();
                 } else if (token == JsonToken.VALUE_NUMBER_FLOAT) {
-                    value = jp.getFloatValue();
+                    value = jp.getDoubleValue();
                 } else if (token == JsonToken.VALUE_TRUE) {
                     value = true;
                 } else if (token == JsonToken.VALUE_FALSE) {
